EDU-1009 Move new media components into main folder
