+++ src/components/media-player/html5-player.less
