+++ src/components/media-player/media-player.less
