+++ src/components/media-player/multitrack-media-player.less
