+++ src/components/resource-picker/resource-preview.js
@@ -1,17 +1,17 @@
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import prettyBytes from 'pretty-bytes';
-import MiniPager from '../mini-pager.js';
 import { message, Tooltip } from 'antd';
+import MiniPager from '../mini-pager.js';
 import Logger from '../../common/logger.js';
 import { useTranslation } from 'react-i18next';
 import LiteralUrlLink from '../literal-url-link.js';
 import { useService } from '../container-context.js';
-import { handleError } from '../../ui/error-helper.js';
 import { useDateFormat } from '../locale-context.js';
+import { handleError } from '../../ui/error-helper.js';
 import mimeTypeHelper from '../../ui/mime-type-helper.js';
+import MediaPlayer from '../media-player/media-player.js';
 import ClientConfig from '../../bootstrap/client-config.js';
-import MediaPlayer from '../media-player/plyr/media-player.js';
 import { getResourceType } from '../../utils/resource-utils.js';
 import FileTextFilledIcon from '../icons/files/file-text-filled-icon.js';
 import { MEDIA_SCREEN_MODE, RESOURCE_TYPE } from '../../domain/constants.js';
