+++ src/plugins/audio/audio-display.js
@@ -4,7 +4,7 @@ import { MEDIA_SCREEN_MODE } from '../../domain/constants.js';
 import { useService } from '../../components/container-context.js';
 import CopyrightNotice from '../../components/copyright-notice.js';
 import { sectionDisplayProps } from '../../ui/default-prop-types.js';
-import MediaPlayer from '../../components/media-player/plyr/media-player.js';
+import MediaPlayer from '../../components/media-player/media-player.js';
 import { getAccessibleUrl, isInternalSourceType } from '../../utils/source-utils.js';
 
 function AudioDisplay({ content }) {
