+++ src/plugins/media-slideshow/media-slideshow-display.js
@@ -7,7 +7,7 @@ import ClientConfig from '../../bootstrap/client-config.js';
 import { useService } from '../../components/container-context.js';
 import CopyrightNotice from '../../components/copyright-notice.js';
 import { sectionDisplayProps } from '../../ui/default-prop-types.js';
-import MediaPlayer from '../../components/media-player/plyr/media-player.js';
+import MediaPlayer from '../../components/media-player/media-player.js';
 import { MEDIA_ASPECT_RATIO, MEDIA_SCREEN_MODE } from '../../domain/constants.js';
 import { getAccessibleUrl, isInternalSourceType } from '../../utils/source-utils.js';
