+++ src/plugins/video/video-display.js
@@ -3,7 +3,7 @@ import ClientConfig from '../../bootstrap/client-config.js';
 import { useService } from '../../components/container-context.js';
 import CopyrightNotice from '../../components/copyright-notice.js';
 import { sectionDisplayProps } from '../../ui/default-prop-types.js';
-import MediaPlayer from '../../components/media-player/plyr/media-player.js';
+import MediaPlayer from '../../components/media-player/media-player.js';
 import { getAccessibleUrl, isInternalSourceType } from '../../utils/source-utils.js';
 
 function VideoDisplay({ content }) {
