+++ src/styles/main.less
@@ -86,18 +86,18 @@
 @import '../components/admin/storage-plans-tab.less';
 @import '../components/resource-picker/utilities.less';
 @import '../components/never-scrolling-text-area.less';
+@import '../components/media-player/html5-player.less';
+@import '../components/media-player/media-player.less';
 @import '../components/localization/language-icon.less';
 @import '../components/pages/complete-registration.less';
 @import '../components/pages/browser-not-supported.less';
+@import '../components/media-player/youtube-player.less';
 @import '../components/resource-picker/files-viewer.less';
 @import '../components/pages/complete-password-reset.less';
-@import '../components/media-player/plyr/html5-player.less';
 @import '../components/pages/connect-external-account.less';
-@import '../components/media-player/plyr/media-player.less';
 @import '../components/media-player/track-mixer-editor.less';
 @import '../components/admin/technical-maintenance-tab.less';
 @import '../components/media-player/track-mixer-display.less';
-@import '../components/media-player/plyr/youtube-player.less';
 @import '../components/external-account-provider-dialog.less';
 @import '../components/resource-picker/resource-preview.less';
 @import '../components/media-player/media-volume-slider.less';
@@ -114,9 +114,9 @@
 @import '../components/pages/room-membership-confirmation.less';
 @import '../components/localization/language-flag-and-name.less';
 @import '../components/resource-picker/files-upload-screen.less';
+@import '../components/media-player/multitrack-media-player.less';
 @import '../components/media-player/media-player-progress-bar.less';
 @import '../components/resource-picker/wikimedia-files-viewer.less';
-@import '../components/media-player/plyr/multitrack-media-player.less';
 
 
 // Pure CSS components
